+++ lib/models/data_io/import_qfx.dart
@@ -86,39 +86,39 @@ class OfxBankInfo {
 int getCategoryFromOfxType(final QFXTransaction ofxTransaction) {
   int categoryId = -1;
   switch (ofxTransaction.type) {
-    case "CREDIT":
+    case 'CREDIT':
       categoryId = Data().categories.investmentCredit.id.value;
       break;
-    case "DEBIT":
+    case 'DEBIT':
       categoryId = Data().categories.investmentDebit.id.value;
       break;
-    case "INT":
+    case 'INT':
       categoryId = Data().categories.investmentInterest.id.value;
       break;
-    case "DIV":
+    case 'DIV':
       categoryId = Data().categories.investmentDividends.id.value;
       break;
-    case "FEE":
-    case "SRVCHG": // service charge
+    case 'FEE':
+    case 'SRVCHG': // service charge
       categoryId = Data().categories.investmentFees.id.value;
       break;
-    case "DEP": // deposit
-    case "ATM": // automatic teller machine
-    case "POS": // Point of sale
-    case "PAYMENT": // Electronic payment
-    case "CASH": // Cash withdrawal
-    case "DIRECTDEP": // Direct deposit
-    case "DIRECTDEBIT": // Direct debit;
-    case "REPEATPMT": // Repeating payment
-    case "CHECK": // check
-    case "OTHER":
+    case 'DEP': // deposit
+    case 'ATM': // automatic teller machine
+    case 'POS': // Point of sale
+    case 'PAYMENT': // Electronic payment
+    case 'CASH': // Cash withdrawal
+    case 'DIRECTDEP': // Direct deposit
+    case 'DIRECTDEBIT': // Direct debit;
+    case 'REPEATPMT': // Repeating payment
+    case 'CHECK': // check
+    case 'OTHER':
       if (ofxTransaction.amount > 0) {
         categoryId = Data().categories.investmentCredit.id.value;
       } else {
         categoryId = Data().categories.investmentDebit.id.value;
       }
       break;
-    case "XFER":
+    case 'XFER':
       categoryId = Data().categories.investmentTransfer.id.value;
       break;
   }